Keep file policy tests within line budget
